requirements.txt
discord
request
BeautifulSoup
pprint
decimal
urllib
collections
hashlib
codecs
flask
datetime
PIL
linebot
